multikomh

+++ include/MinimizerGenerator.hpp
@@ -20,9 +20,9 @@ public:
     MinimizerGenerator(std::vector<std::vector<seqan3::dna5>> unique_reads, cmd_arguments args);
     // MinimizerGenerator(std::map<std::vector<seqan3::dna5>, uint32_t> read2count, cmd_arguments args);
     std::unordered_map<std::uint64_t, std::vector<std::vector<seqan3::dna5>>> minimizer2reads_main();
-    double proba(int L, int k);
+    double proba(unsigned L, unsigned k);
     int kSize(int L, double p);
-    std::tuple<int, int, int, double> possibleBetterParameters(int L, int dt, double pt);
+    std::tuple<unsigned, unsigned, unsigned, double> possibleBetterParameters(unsigned L, uint8_t dt, double pt);
     // std::tuple<int, int, double> findBestParameters(int l, int dt, double pt);
     // long double prob(int l, int n, int k, int dt);
     // void process_read(const std::vector<seqan3::dna5> &read);
